ai-bi-dashboard/requirements.txt
streamlit
pandas
matplotlib
langchain
langchain-google-genai
langchain-experimental
python-dotenv
tabulate
plotly
google-generativeai
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pydantic[email]